Move chip stage utils to separate file for organization
