+++ utils/src/index.ts
@@ -49,6 +49,7 @@ export * from './stages/chat_stage.prompts';
 export * from './stages/chat_stage.validation';
 export * from './stages/chip_stage';
 export * from './stages/chip_stage.prompts';
+export * from './stages/chip_stage.utils';
 export * from './stages/chip_stage.validation';
 export * from './stages/comprehension_stage';
 export * from './stages/comprehension_stage.validation';
